% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_architecture.R
\name{mod_architecture_ui}
\alias{mod_architecture_ui}
\alias{mod_architecture_server}
\title{mod_architecture_ui and mod_architecture_server}
\usage{
mod_architecture_ui(id)

mod_architecture_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
