% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_future.R
\name{mod_future_ui}
\alias{mod_future_ui}
\alias{mod_future_server}
\title{mod_future_ui and mod_future_server}
\usage{
mod_future_ui(id)

mod_future_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
