% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shiny_intro.R
\name{mod_shiny_intro_ui}
\alias{mod_shiny_intro_ui}
\alias{mod_shiny_intro_server}
\title{mod_shiny_intro_ui and mod_shiny_intro_server}
\usage{
mod_shiny_intro_ui(id)

mod_shiny_intro_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
