% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ui_ux.R
\name{mod_ui_ux_ui}
\alias{mod_ui_ux_ui}
\alias{mod_ui_ux_server}
\title{mod_ui_ux_ui and mod_ui_ux_server}
\usage{
mod_ui_ux_ui(id)

mod_ui_ux_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
