% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shiny_rev_db}
\alias{shiny_rev_db}
\title{Summary data of shiny-related packages on CRAN}
\format{A data frame with the following columns:
\itemize{
\item \code{package}: name of package
\item \code{title}: Package title
\item \code{description}: Package description
\item \code{author}: List column of the package author(s)
\item \code{published}: Date for when package was last published to CRAN
\item \code{n_releases}: Number of releases on CRAN
\item \code{first_release_date}: Date when package was first published to CRAN
\item \code{last_release_date}: Date when package was last published to CRAN
}}
\usage{
shiny_rev_db
}
\description{
A dataset containing metadata for packages on CRAN that import \code{shiny}
but are not authored by RStudio.
}
\note{
see \code{data-raw/shiny_rev_db.R} script for how this was generated
}
\keyword{datasets}
