% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_background.R
\name{title_background}
\alias{title_background}
\title{Create background for title page}
\usage{
title_background(
  hex_path = system.file("app", "www", "img", package = "highlights.shiny"),
  background_file = "blueprint.jpg",
  sticker_width = 200,
  sticker_row_size = 5,
  hex_scale_pct_width = 50,
  hex_scale_pct_height = NULL,
  hex_offset_vec = c(500, 200),
  brightness = 250,
  saturation = 140,
  hue = 100,
  radius = 10,
  sigma = 5
)
}
\arguments{
\item{hex_path}{directory for where hex png files are stored}

\item{background_file}{image file for background of slide}

\item{sticker_width}{width of each sticker (in pixels)}

\item{sticker_row_size}{number of rows for the stickers. Default is 5.}

\item{brightness}{brightness of background supplied to \code{magick::image_modulate}}

\item{saturation}{saturation of background supplied to \code{magick::image_modulate}}

\item{hue}{hue of background supplied to \code{magick::image_modulate}}

\item{radius}{radius of background blur supplied to \code{magick::image_blur}}

\item{sigma}{sigma of background blur supplied to \code{image_blur}}
}
\value{
image produced by \code{magick::image_composite}
}
\description{
Create background for title page
}
