% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_background.R
\name{title_background}
\alias{title_background}
\title{Create background for title page}
\usage{
title_background(
  hex_path = system.file("app", "www", "img", package = "highlights.shiny"),
  background_path = system.file("app", "www", "source_img", package =
    "highlights.shiny"),
  background_file = "sky.jpg",
  sticker_width = 200,
  sticker_row_size = 4,
  hex_scale_pct_width = 50,
  hex_scale_pct_height = NULL,
  hex_offset_vec = c(500, 200),
  background_scale_pct_width = 100,
  background_scale_pct_height = NULL,
  background_offset_vec = c(100, 300),
  brightness = 400,
  saturation = 200,
  hue = 100,
  radius = 10,
  sigma = 5,
  canvas_color = "white"
)
}
\arguments{
\item{hex_path}{directory for where hex png files are stored}

\item{background_file}{image file for background of slide.
If NULL, no background image file will be used.}

\item{sticker_width}{width of each sticker (in pixels)}

\item{sticker_row_size}{number of rows for the stickers. Default is 5.}

\item{brightness}{brightness of background supplied to \code{magick::image_modulate}}

\item{saturation}{saturation of background supplied to \code{magick::image_modulate}}

\item{hue}{hue of background supplied to \code{magick::image_modulate}}

\item{radius}{radius of background blur supplied to \code{magick::image_blur}}

\item{sigma}{sigma of background blur supplied to \code{image_blur}}
}
\value{
image produced by \code{magick::image_composite}
}
\description{
Create background for title page
}
